requirements.txt
pycryptodome
requests
browser-cookie3
